[Gemspec] - Allow lower faraday versions (#173)
# Description

In reference to issue: https://github.com/crmne/ruby_llm/issues/172 -
this CR exists to allow for applications requiring different, lower
Faraday versions to integrate with this Gem.

# Changes Made

1. Updated the gemspec for Faraday and its dependencies to allow use of
lower versions in case of a conflict.
2. Updated CICD.yml workflow to validate that lower Faraday versions
work.
3. Specs failed on lower Faraday versions so I updated
`lib/ruby_llm/streaming.rb` to allow for legacy Faraday versions to
stream responses as well. Could use feedback definitely in case this
breaks logic in a way you all are uncomfy with.

# Testing

I ran manually rspec with Faraday Version 2.13.1 and 1.10.3, here are
the results prior to CI/CD integration in this repo:

| Faraday Version   |    Validate Gem Install        |  Test Results |
|----------|:-------------:|------:|
| 1.10.3 | ![Screenshot 2025-05-14 at 12 52
57 PM](https://github.com/user-attachments/assets/5233a968-c37f-4de3-9ef0-bdaad17a3a2e)
| ![Screenshot 2025-05-14 at 12 53
03 PM](https://github.com/user-attachments/assets/b21f830f-732a-4aef-94fa-5db07a2be8f8)
|
| 2.13.1 | ![Screenshot 2025-05-14 at 12 52
39 PM](https://github.com/user-attachments/assets/78c81a09-0bed-4f10-bc72-c8e76d16df08)
| ![Screenshot 2025-05-14 at 12 52
46 PM](https://github.com/user-attachments/assets/ded40e55-7296-49a8-b4c8-b11a49ad6726)
|